requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.13
django-4-jet==1.0.8
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-countries==7.3.2
django-countries-with-calling-codes==0.1.1
django-custom-logging==0.1.1
django-filter==21.1
django-jet==1.0.0
django-modeltranslation==0.17.7
django-nested-admin==3.4.0
django-timezone-field==4.2.3
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
engineering-notation==0.6.0
frozenlist==1.3.0
geoip2==4.5.0
idna==3.3
importlib-metadata==4.11.3
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.3.7
MarkupSafe==2.1.1
maxminddb==2.2.0
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-monkey-business==1.0.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
text-unidecode==1.3
tk-tools==0.14.0
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
